third_party/blink/web_tests/svg/parser/whitespace-length-expected.txt

CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47 em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1 em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35 em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10 em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32 em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1 em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2 em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47 ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1 ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35 ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10 ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32 ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1 ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2 ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47 px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1 px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35 px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10 px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32 px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1 px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2 px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47 in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1 in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35 in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10 in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32 in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1 in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2 in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47 cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1 cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35 cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10 cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32 cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1 cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2 cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47 mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1 mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35 mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10 mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32 mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1 mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2 mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47 pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1 pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35 pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10 pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32 pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1 pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2 pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47 pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1 pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35 pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10 pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32 pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1 pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2 pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47 %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1 %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35 %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10 %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32 %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1 %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2 %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -47a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .1a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 0.35a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 1e-10a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +32a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +17E-1a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 17e+2a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -47e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .1e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 0.35e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 1e-10e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +32e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +17E-1e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 17e+2e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -47foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .1foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 0.35foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 1e-10foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +32foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +17E-1foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 17e+2foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -47)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .1)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 0.35)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 1e-10)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +32)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +17E-1)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " 17e+2)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47   em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1   em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35   em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10   em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32   em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1   em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2   em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47   ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1   ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35   ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10   ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32   ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1   ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2   ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47   px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1   px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35   px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10   px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32   px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1   px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2   px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47   in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1   in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35   in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10   in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32   in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1   in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2   in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47   cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1   cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35   cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10   cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32   cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1   cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2   cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47   mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1   mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35   mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10   mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32   mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1   mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2   mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47   pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1   pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35   pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10   pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32   pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1   pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2   pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47   pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1   pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35   pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10   pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32   pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1   pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2   pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47   %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1   %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35   %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10   %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32   %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1   %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2   %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -47a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .1a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   0.35a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   1e-10a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +32a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +17E-1a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   17e+2a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -47e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .1e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   0.35e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   1e-10e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +32e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +17E-1e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   17e+2e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -47foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .1foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   0.35foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   1e-10foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +32foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +17E-1foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   17e+2foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -47)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .1)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   0.35)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   1e-10)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +32)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +17E-1)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   17e+2)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\r\n\t em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\r\n\t em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\r\n\t em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\r\n\t em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\r\n\t em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\r\n\t em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\r\n\t em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\r\n\t ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\r\n\t ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\r\n\t ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\r\n\t ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\r\n\t ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\r\n\t ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\r\n\t ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\r\n\t px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\r\n\t px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\r\n\t px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\r\n\t px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\r\n\t px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\r\n\t px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\r\n\t px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\r\n\t in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\r\n\t in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\r\n\t in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\r\n\t in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\r\n\t in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\r\n\t in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\r\n\t in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\r\n\t cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\r\n\t cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\r\n\t cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\r\n\t cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\r\n\t cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\r\n\t cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\r\n\t cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\r\n\t mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\r\n\t mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\r\n\t mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\r\n\t mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\r\n\t mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\r\n\t mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\r\n\t mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\r\n\t pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\r\n\t pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\r\n\t pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\r\n\t pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\r\n\t pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\r\n\t pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\r\n\t pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\r\n\t pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\r\n\t pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\r\n\t pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\r\n\t pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\r\n\t pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\r\n\t pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\r\n\t pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\r\n\t %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\r\n\t %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\r\n\t %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\r\n\t %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\r\n\t %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\r\n\t %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\r\n\t %".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -47a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .1a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 0.35a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 1e-10a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +32a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +17E-1a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 17e+2a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -47e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .1e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 0.35e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 1e-10e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +32e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +17E-1e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 17e+2e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -47foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .1foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 0.35foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 1e-10foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +32foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +17E-1foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 17e+2foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -47)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .1)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 0.35)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 1e-10)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +32)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +17E-1)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t 17e+2)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\fem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\fem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\fem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\fem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\fem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\fem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\fem".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\fex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\fex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\fex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\fex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\fex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\fex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\fex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\fpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\fpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\fpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\fpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\fpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\fpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\fpx".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\fin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\fin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\fin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\fin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\fin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\fin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\fin".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\fcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\fcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\fcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\fcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\fcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\fcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\fcm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\fmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\fmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\fmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\fmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\fmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\fmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\fmm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\fpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\fpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\fpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\fpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\fpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\fpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\fpt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\fpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\fpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\fpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\fpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\fpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\fpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\fpc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-47\f%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".1\f%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "0.35\f%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "1e-10\f%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+32\f%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+17E-1\f%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "17e+2\f%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-47a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.1a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f0.35a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f1e-10a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+32a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+17E-1a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f17e+2a".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-47e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.1e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f0.35e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f1e-10e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+32e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+17E-1e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f17e+2e".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-47foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.1foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f0.35foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f1e-10foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+32foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+17E-1foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f17e+2foo".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-47)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.1)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f0.35)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f1e-10)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+32)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+17E-1)90".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f17e+2)90".
This is a testharness.js-based test.
Harness: the test ran to completion.
